requirements.txt
os.path
re
lxml.etree
requests